requirements.txt
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
constantly==23.10.4
cryptography==41.0.5
daphne==4.0.0
dj-database-url==2.1.0
Django==5.0
django-cors-headers==4.2.0
django-extensions==3.2.3
django-filter==23.3
django-jazzmin==2.6.0
django-js-asset==2.1.0
django-mptt==0.15.0
django-suit @ https://github.com/darklow/django-suit/tarball/v2#sha256=b40c916de9528ca86bcccd42d1fd54e7773e5071d2bfd9efc576039c1a441c0b
django-taggit==5.0.1
django-unfold==0.21.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
et-xmlfile==1.1.0
gunicorn==21.2.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
msgpack==1.0.7
openpyxl==3.1.2
packaging==23.2
Pillow==10.0.1
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydot==2.0.0
PyJWT==2.8.0
pyOpenSSL==23.3.0
pyparsing==3.1.1
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
referencing==0.34.0
rpds-py==0.18.0
service-identity==23.1.0
setuptools==69.5.1
six==1.16.0
sqlparse==0.4.4
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.8.0
uritemplate==4.1.1
whitenoise==6.6.0
zope.interface==6.1
